### 2025-03-03 #7865 @editorialbot
[PRE REVIEW]: LazyModeler: An R package for automatic simplification, check, and visualization of regression models
**Submitting author:** @LMKoesters (<a href="http://orcid.org/0000-0002-7913-2377">Lara M Kösters</a>)
**Repository:** https://github.com/LMKoesters/LazyModeler
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.0
