### 2025-04-28 #7865 @fabian-s
@bbortey9  @salbalkus @gavinsimpson

would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
Please be so kind to reply even if you are not interested, so that I can start looking for other reviewers.
