### 2025-04-28 #7865 @gavinsimpson
@fabian-s yeah, I can do a review of the package for JOSS
