### 2025-04-28 #7865 @fabian-s
tysm @gavinsimpson
