### 2025-04-28 #8127 @editorialbot
[REVIEW]: LazyModeler: An R package for automatic simplification, check, and visualization of regression models
**Submitting author:** @LMKoesters (<a href="http://orcid.org/0000-0002-7913-2377">Lara M Kösters</a>)
**Repository:** https://github.com/LMKoesters/LazyModeler
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.0
