### 2025-04-28 #8127 @fabian-s
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> ❌ MISSING DOIs
>
> - 10.32614/cran.package.easystats may be a valid DOI for title: easystats: Framework for Easy Statistical Modeling...
> - 10.32614/cran.package.mumin may be a valid DOI for title: MuMIn: Multi-Model Inference
> - None
> ```

 @LMKoesters : while we wait for reviewer feedback, please check these missing DOIs and add them if correct
