### 2025-04-29 #7865 @bbortey9
@fabian-s <https://github.com/fabian-s> , yes I can do the review

On Mon, Apr 28, 2025 at 8:07 AM Fabian Scheipl ***@***.***>
wrote:

> *fabian-s* left a comment (openjournals/joss-reviews#7865)
> <https://github.com/openjournals/joss-reviews/issues/7865#issuecomment-2835029464>
>
> @bbortey9 <https://github.com/bbortey9> @salbalkus
> <https://github.com/salbalkus> @gavinsimpson
> <https://github.com/gavinsimpson>
>
> would any of you be willing to review this submission for JOSS? We carry
> out our checklist-driven reviews here in GitHub issues and follow these
> guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
> Please be so kind to reply even if you are not interested, so that I can
> start looking for other reviewers.
>
