### 2025-04-29 #7865 @fabian-s
@bbortey9 thanks, we've already reached the number of required referees. Sorry to have bothered you with this, and we do appreciate your quick response. Hopefully, JOSS can rely on your expertise another time.
