### 2025-04-29 #7865 @bbortey9
No problem, I understand.

On Tue, Apr 29, 2025 at 3:49 AM Fabian Scheipl ***@***.***>
wrote:

> *fabian-s* left a comment (openjournals/joss-reviews#7865)
> <https://github.com/openjournals/joss-reviews/issues/7865#issuecomment-2837835381>
>
> @bbortey9 <https://github.com/bbortey9> thanks, we've already reached the
> number of required referees. Sorry to have bothered you with this, and we
> do appreciate your quick response. Hopefully, JOSS can rely on your
> expertise another time.
>
