### 2025-05-05 8e469d0
adjusted DOIs in bib
