+++ paper/paper.bib
@@ -267,15 +267,6 @@ year = {2024},
 url = {http://www.posit.co/},
 }
 
-@article{Luedecke2024,
-title = {easystats: Framework for Easy Statistical Modeling, Visualization, and Reporting},
-author = {Daniel Lüdecke and Mattan S. Ben-Shachar and Indrajeet Patil and Brenton M. Wiernik and Etienne Bacher and Rémi Thériault and Dominique Makowski},
-journal = {CRAN},
-year = {2022},
-note = {R package},
-url = {https://easystats.github.io/easystats/},
-}
-
 @article{Luedecke2021, 
 doi = {10.21105/joss.03139}, 
 url = {https://doi.org/10.21105/joss.03139}, 
@@ -476,6 +467,7 @@ title = {MuMIn: Multi-Model Inference},
 author = {Kamil Bartoń},
 year = {2024},
 note = {R package version 1.48.4},
+doi = {10.32614/cran.package.mumin},
 url = {https://CRAN.R-project.org/package=MuMIn},
 }
 
