### 2025-05-20 LMKoesters/LazyModeler#1 @strengejacke
[JOSS review] Installation instruction missing
related to https://github.com/openjournals/joss-reviews/issues/8127

The readme could have a very short paragraph / sentence in how to install the package (via CRAN and/or GitHub).
