### 2025-05-20 LMKoesters/LazyModeler#2 @strengejacke
[JOSS review] Statement of Need in readme
Related to https://github.com/openjournals/joss-reviews/issues/8127

You could maybe add one or two sentences in the readme that shows why this software is needed and how it compares to other packages (similar to what you wrote in the paper). It doesn't have to be a full statement of need sections, a few additional details on that topic in the readme here on GitHub would be sufficient.
