### 2025-05-20 LMKoesters/LazyModeler#3 @strengejacke
[JOSS review] Community Guidelines
Related to https://github.com/openjournals/joss-reviews/issues/8127

The readme lacks of a section on Community Guidlines / how to contribute / how to find or seek help.
