### 2025-05-20 LMKoesters/LazyModeler#4 @strengejacke
[JOSS review] Conclusion section in paper
Related to https://github.com/openjournals/joss-reviews/issues/8127

Would be nice to have a short conclusion paragraph at the end in the paper.
