### 2025-06-12 #8127 @strengejacke
In my case, I cannot finalize the reviewer feedback until https://github.com/LMKoesters/LazyModeler/issues/1 is resolved, because I need to state whether _"Does installation proceed as outlined in the documentation?"_. I suggest addressing the issues related to this review, so we / I can approve the manuscript and software submission.
