### 2025-07-08 57c03aa
added examples; renamed not_so_expanded_model_summary to slightly_expanded_model_summary; added to README
