+++ R/LazyModeler.R
@@ -29,20 +29,20 @@ remove_autocorrelations <- function(df, coefficients, ..., automatic_removal=TRU
   
   correlations_l = correlations %>%
     rownames_to_column(var='coefficientA') %>%
-    pivot_longer(!.data$coefficientA, names_to = "coefficientB", values_to = "correlation")
+    pivot_longer(!coefficientA, names_to = "coefficientB", values_to = "correlation")
   
   correlations_p_values_l = correlation_p_values %>%
     rownames_to_column(var='coefficientA') %>%
-    pivot_longer(!.data$coefficientA, names_to = "coefficientB", values_to = "p_value")
+    pivot_longer(!coefficientA, names_to = "coefficientB", values_to = "p_value")
   
   correlations_complete = merge(correlations_l, correlations_p_values_l, by=c('coefficientA', 'coefficientB'))
   
   correlations_complete = correlations_complete %>%
-    filter(.data$coefficientA != .data$coefficientB) %>%
-    mutate(col1 = pmin(.data$coefficientA, .data$coefficientB), col2 = pmax(.data$coefficientA, .data$coefficientB), 
-           comparison = paste(.data$col1, .data$col2)) %>%
-    distinct(.data$comparison, .keep_all = TRUE) %>%
-    select(!c(.data$col1, .data$col2, .data$comparison))
+    filter(coefficientA != coefficientB) %>%
+    mutate(col1 = pmin(coefficientA, coefficientB), col2 = pmax(coefficientA, coefficientB), 
+           comparison = paste(col1, col2)) %>%
+    distinct(comparison, .keep_all = TRUE) %>%
+    select(!c(col1, col2, comparison))
   
   autocorrelations = correlations_complete[((correlations_complete$cor >= autocorrelation_threshold) | 
                                                (correlations_complete$cor <= -autocorrelation_threshold)) & 
@@ -59,10 +59,10 @@ remove_autocorrelations <- function(df, coefficients, ..., automatic_removal=TRU
       autocorrelations = merge(autocorrelations, coefficients_df[c('coefficient', 'idx')], by.x='coefficientB', by.y='coefficient', suffixes = c('1', '2'))
       autocorrelations = autocorrelations %>%
         rowwise() %>%
-        mutate(idx_smaller = sort(c(.data$idx1, .data$idx2))[[1]],
-               idx_bigger = sort(c(.data$idx1, .data$idx2))[[2]],
-               comb = paste(.data$idx_smaller, .data$idx_bigger, sep='|')) %>%
-        arrange(desc(.data$idx_bigger))
+        mutate(idx_smaller = sort(c(idx1, idx2))[[1]],
+               idx_bigger = sort(c(idx1, idx2))[[2]],
+               comb = paste(idx_smaller, idx_bigger, sep='|')) %>%
+        arrange(desc(idx_bigger))
       
       for (i in 1:nrow(autocorrelations)) {
         # C is the least important and part of comparison
@@ -118,6 +118,11 @@ remove_autocorrelations <- function(df, coefficients, ..., automatic_removal=TRU
 #' @param categorical_vars List of categorical variables within the model (base names, i.e., names of columns)
 #' @param col Column to be used for check. Default is 'coefficients'
 #' @return Language object with if/else statements to be applied to model summary with information on whether model variables are categorical
+#' @examples
+#' data("plants");
+#' 
+#' setup_categorical_check(plants, c('habitat', 'ploidy'), col=quote(predictor))
+#' @export
 setup_categorical_check <- function(df, categorical_vars, col=quote(coefficients)) {
   if (length(categorical_vars) > 0) {
     categorical_check = c()
[…]
