### 2025-07-08 8f3459d
adjusted plot_model_features example
