### 2025-09-01 c1c3382
added examples to some now exported functions; adjusted citations in paper and added, e.g., installation instructions to README (fixes #1, fixes #2, fixes #3, fixes #4, fixes #5)
