+++ paper/paper.bib
@@ -58,7 +58,7 @@ journal = {Ecology and Evolution},
 month = {dec},
 number = {24},
 pages = {14066--14084},
-title = {{Desert‐like badlands and surrounding (semi‐)dry grasslands of Central Germany promote small‐scale phenotypic and genetic differentiation in Thymus praecox}},
+title = {{Desert‐like badlands and surrounding (semi‐)dry grasslands of Central Germany promote small‐scale phenotypic and genetic differentiation in {\emph{Thymus praecox}}}},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/ece3.5844},
 volume = {9},
 year = {2019}
@@ -369,11 +369,11 @@ doi = {10.1029/2021JG006238},
 file = {:Users/kevin/Downloads/Liaw2021_Explaining_Shortcomings_Log-Transforming_Dependent_Variable.pdf:pdf},
 issn = {2169-8953},
 journal = {Journal of Geophysical Research: Biogeosciences},
-keywords = {Gaussian-Gamma model,Mediterranean climate,carbon cycle,desert,forest,grassland,log-transforming dependent variable,marsh,nonlinear estimation method,soil CO2 emissions,soil respiration,soil temperature,soil water content,temperate climate,under-prediction,wetland},
+keywords = {Gaussian-Gamma model,Mediterranean climate,carbon cycle,desert,forest,grassland,log-transforming dependent variable,marsh,nonlinear estimation method,soil CO$_2$ emissions,soil respiration,soil temperature,soil water content,temperate climate,under-prediction,wetland},
 month = {may},
 number = {5},
 pages = {1--18},
-title = {{Explaining the shortcomings of log‐transforming the dependent variable in regression models and recommending a better alternative: Evidence from soil CO2 emission studies}},
+title = {{Explaining the shortcomings of log‐transforming the dependent variable in regression models and recommending a better alternative: Evidence from soil CO$_2$ emission studies}},
 url = {https://agupubs.onlinelibrary.wiley.com/doi/10.1029/2021JG006238},
 volume = {126},
 year = {2021}
