### 2025-09-01 0394c58
Increment version number to 0.2.2
