+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: LazyModeler
 Type: Package
 Title: LazyModeler: An R package for automatic simplification, check, and visualization of regression models
-Version: 0.2.1
+Version: 0.2.2
 Authors@R: 
     c(person(given = "Lara",
              family = "Koesters",
