### 2025-09-28 97f741f
Update installation instructions in README.md
This provides more standard installation instructions:

1. it is a package, not a library - the latter is where packages are installed,
2. use the *remotes* package directly to install, *devtools* just re-exports `remotes::install_github()` and loads a lot of stuff a user doesn't need just to use your package
3. to that same end, don't load and attach *remotes*, just load it by calling `remotes::install_github()` instead of calling `library()` then `install_github()`.
