### 2025-09-28 LMKoesters/LazyModeler#7 @gavinsimpson
[JOSS Review] Missing dependency on R versions
Related to openjournals/joss-reviews#8127

When I install the current GitHub version I see the following when I follow the installation instructions:

```
── R CMD build ──────────────────────────────────────────────────────────────────────────────────────────────────
   checking for file ‘/private/var/folders/yq/gqj7yv7j3n17w__256r9vxn40000gn/T/RtmpNOJQfF/remotes2a2469fb742/LMKo✔  checking for file ‘/private/var/folders/yq/gqj7yv7j3n17w__256r9vxn40000gn/T/RtmpNOJQfF/remotes2a2469fb742/LMKoesters-LazyModeler-03c845f/DESCRIPTION’
─  preparing ‘LazyModeler’:
     NB: this package now depends on R (>= 3.5.0)
     WARNING: Added dependency on R >= 3.5.0 because serialized objects in
     serialize/load version 3 cannot be read in older versions of R.
     File(s) containing such objects:
       ‘LazyModeler/data/plants.rda’
     NB: this package now depends on R (>= 4.1.0)
     WARNING: Added dependency on R >= 4.1.0 because package code uses the
     pipe |> or function shorthand \(...) syntax added in R 4.1.0.
     File(s) using such syntax:
       ‘plot_model_features.Rd’
─  building ‘LazyModeler_0.2.2.tar.gz’
```

As stated, you need to depend on R >= 4.1.0 because you use the native pipe `|>` in the documentation (roxygen) for the `plot_model_features()` function.

You could avoid this by switching to the *magritte* pipe that you seem to use in the package code (and by loading the required package to allow use of `%>%` in the example), but even then you'll need to add a dependency on R >= 3.5.0 because of the version of the `.rda` file used for the data set stored in the package.

## Solution

Add either

```
Depends: R (>= 4.1.0)
```

to the `DESCRIPTION`, or fix the example to use the magritte pipe.
