### 2025-09-28 #8127 @gavinsimpson
@fabian-s I have now completed my review. I have opened several issues for specific problems, which are supported by a longer narrative review. I have also submitted a PR to fix an issue with the installation instructions.
