### 2025-09-30 8e3566d
Update paper.md
I addressed the first three issues of [JOSS Review] Comments on paper #9. Please check.
