+++ paper/paper.md
@@ -8,7 +8,7 @@ tags:
 - (non)linear regression models (LM, GLM, LMER, GLMER, GAM, or NLMER)
 - relationship plotting
 
-date: "25 February 2025"
+date: "30 September 2025"
 output: pdf_document
 authors:
 - name: Lara M. Kösters
@@ -29,13 +29,13 @@ affiliations:
 
 # Summary
 
-Setting up, simplifying, checking, and visualizing regression models continues to be a time-consuming task involving multiple, sometimes concurrent, workflows and software packages. This particularly applies to big data research where several models need to be set up and optimized. To tackle this problem, we present `LazyModeler` - a statistical package for the programming language R that allows to easily perform regression modeling. It includes removal of autocorrelated variables, choice between several types of (non)linear regression models, standard stepwise model simplification, various model quality checks, plotting of coefficient estimates and relationships, and output generation. `LazyModeler` will significantly speed up regression modeling, enabling people to analyze and illustrate their data in a statistically reliable and standardized manner.
+Setting up, simplifying, checking, and visualizing regression models continues to be a time-consuming task involving multiple, sometimes concurrent, workflows and software packages. This particularly applies to big data research where several models with different resonse variables and many explanatory variables need to be set up and optimized. To tackle this problem, we present `LazyModeler` - a statistical package for the programming language R that allows to easily perform regression modeling. It includes removal of autocorrelated variables, choice between several types of (non)linear regression models, standard stepwise model simplification, various model quality checks, plotting of coefficient estimates and relationships, and output generation. `LazyModeler` will significantly speed up regression modeling, enabling people to analyze and illustrate their data in a statistically reliable and standardized manner.
 
 # Statement of need
 
 Statistical modeling describes the process of finding a mathematical function with specific statistical assumptions that best fits the observed data [@Crawley2007; @Crawley2015; @Henley2020]. This process attempts, in practice, to find a (causal) relationship between a dependent response variable `y` and an independent predictor variable `x` for any postulated hypothesis. For statistical inference and graphics in science, the programming environment R [@RCoreTeam2024] has become highly popular.
 
-Linear regression models, as one of the most basic and powerful tools, have been frequently applied in this context [@Crawley2007; @Crawley2015; @Schielzeth2020; @Li2023]. Because of their flexibility, they also allow for non-normally distributed response variables (e.g., in the case of binomial, proportional, or count data), and any kind of transformation for numerical (e.g., polynomial or logarithmic) and categorical (e.g., centered or one-hot/fractional encoded) predictor variables, as well as interactions among them [@Cai2023; @Roemermann2016; @Henley2020; @Karbstein2019; @Karbstein2020; @Karbstein2021; @Liaw2021; @Schielzeth2010]. Regression models also provide the ability to control for random effects that may influence the variables of interest [e.g., @Bauer2020; @Wicke2016; @Schielzeth2020]. Although other statistical technologies can outperform them in highly complex, non-linear scenarios, regression models allow for detailed variable transformation and interaction, mathematical formula specification, calculation of effect sizes, determination of variable significance, and thus hypothesis testing and explanation [@Benjamin2018; @Bzdok2019; @Schulz2020; @Cai2023; @Karbstein2023; @Li2023]. Recent developments make regression models also applicable to nonlinear scenarios [e.g., @Bates2024; @Hastie2023]. Consequently, they are of high practical value in finding and interpreting significant relationships.
[…]
