### 2025-09-30 9a123ab
Update paper.md
Type in correction to [JOSS Review] Comments on paper #9
