+++ paper/paper.md
@@ -29,7 +29,7 @@ affiliations:
 
 # Summary
 
-Setting up, simplifying, checking, and visualizing regression models continues to be a time-consuming task involving multiple, sometimes concurrent, workflows and software packages. This particularly applies to big data research where several models with different resonse variables and many explanatory variables need to be set up and optimized. To tackle this problem, we present `LazyModeler` - a statistical package for the programming language R that allows to easily perform regression modeling. It includes removal of autocorrelated variables, choice between several types of (non)linear regression models, standard stepwise model simplification, various model quality checks, plotting of coefficient estimates and relationships, and output generation. `LazyModeler` will significantly speed up regression modeling, enabling people to analyze and illustrate their data in a statistically reliable and standardized manner.
+Setting up, simplifying, checking, and visualizing regression models continues to be a time-consuming task involving multiple, sometimes concurrent, workflows and software packages. This particularly applies to big data research where several models with different response variables and many explanatory variables need to be set up and optimized. To tackle this problem, we present `LazyModeler` - a statistical package for the programming language R that allows to easily perform regression modeling. It includes removal of autocorrelated variables, choice between several types of (non)linear regression models, standard stepwise model simplification, various model quality checks, plotting of coefficient estimates and relationships, and output generation. `LazyModeler` will significantly speed up regression modeling, enabling people to analyze and illustrate their data in a statistically reliable and standardized manner.
 
 # Statement of need
 
