### 2025-09-30 e3b2028
Update README.md
I addressed the Philosophical Disagreement Statement of [JOSS Review] Comments on paper #9
