### 2025-09-30 a23bb49
Update paper.md
I addressed the Philosophical Disagreement statement of [JOSS Review] Comments on paper #9
