### 2025-09-30 7873ee1
Update paper.bib
