+++ paper/paper.bib
@@ -482,3 +482,14 @@ number = {13},
 pages = {1--26},
 doi = {10.18637/jss.v082.i13},
 }
+
+@article{Lee2016,
+  title={Exact post-selection inference, with application to the lasso},
+  author={Lee, J. D and Sun, D. L. and Sun, Y. and Taylor, J. E.},
+journal = {Annals of Statistics},
+year = {2016},
+volume = {44},
+number = {3},
+pages = {907--927},
+doi = {10.1214/15-AOS1371},
+}
