### 2025-09-30 87f4abd
Update paper.md
