### 2025-09-30 c299933
Update paper.md
