### 2025-09-30 40b0315
Update paper.md
