+++ paper/paper.bib
@@ -485,7 +485,7 @@ doi = {10.18637/jss.v082.i13},
 
 @article{Lee2016,
   title={Exact post-selection inference, with application to the lasso},
-  author={Lee, J. D and Sun, D. L. and Sun, Y. and Taylor, J. E.},
+  author={Lee, Jason D and Sun, Dennis L. and Sun, Yuekai and Taylor, Jonathan E.},
 journal = {Annals of Statistics},
 year = {2016},
 volume = {44},
