### 2025-10-13 5d86d17
Update README.md
Edited Important Note for [JOSS Review] Comments on paper #9
