### 2025-10-13 f9e92ee
Update paper.md
Edited Important Note for [JOSS Review] Comments on paper #9
