### 2025-10-13 78d1fd4
Update paper.bib
Added citation Selcorr for [JOSS Review] Comments on paper #9
