+++ paper/paper.bib
@@ -493,3 +493,11 @@ number = {3},
 pages = {907--927},
 doi = {10.1214/15-AOS1371},
 }
+
+@Manual{,
+    title = {selcorr: Post-Selection Inference for Generalized Linear Models},
+    author = {Marco Cattaneo},
+    year = {2021},
+    note = {R package version 1.0},
+    url = {https://CRAN.R-project.org/package=selcorr},
+  }
