### 2025-10-13 0b7aaa2
Update paper.bib
fix references
