+++ paper/paper.bib
@@ -494,8 +494,8 @@ pages = {907--927},
 doi = {10.1214/15-AOS1371},
 }
 
-@Manual{,
-    title = {selcorr: Post-Selection Inference for Generalized Linear Models},
+@Manual{Cattaneo2021,
+    title = {selcorr: post-selection inference for generalized linear models},
     author = {Marco Cattaneo},
     year = {2021},
     note = {R package version 1.0},
