### 2025-10-13 4234036
Update README.md
Arrange Code lines for [JOSS Review] Comments on paper #9
