### 2025-10-13 0b9ae80
Update paper.md
Edit Example Code for [JOSS Review] Comments on paper #9
