### 2026-04-08 #8127 @kevinkarbstein
@FabianS, please don't close this yet, as we're still working on resolving all the issues (we've almost solved all of them). Final programming tests are currently being run by @LMKoesters, so please reconsider your decision. Thanks!
