### 2026-04-08 #8127 @LMKoesters
@fabian-s I am truly sorry that we have not yet officially closed all of the issues.
We have incorporated all of the feedback and were basically finished but then encountered a bug that was independent of the review. I have not had the time to properly fix it as I just transitioned jobs and am in the final phase of my PhD. I did not want to inform you that all of the issues were closed before fixing the aforementioned bug. The issue regarding testing is not yet closed as I wanted to check for coverage again after fixing the bug. In its current state, the coverage is >70%.
If you are okay with granting us a couple days more, I will fix the bug this week.
