### 2026-04-27 #8127 @fabian-s
See https://github.com/LMKoesters/LazyModeler/issues/12
