### 2026-05-27 #8127 @kevinkarbstein
@fabian-s Thank you again for the detailed feedback and the time invested in reviewing our package. We appreciate the effort that has gone into the review process.

However, we would like to express our concern regarding the current evaluation outcome, particularly given the substantial revisions we implemented within the requested timeframe. We addressed the previously raised issues and significantly improved the codebase and package structure following the reviewer recommendations.

Importantly, the package functions for its intended purpose, including the examples provided in the documentation. We fully acknowledge that additional improvements and refinements are still possible, and we are committed to addressing the newly identified issues as well. However, we believe that these points mainly concern secondary aspects of the package rather than its core functionality.

We also feel that the statement suggesting that “more issues may appear upon deeper inspection” is difficult to use as a decisive evaluation criterion, since continued refinement is a normal aspect of essentially all actively developed software packages.

@fabian-s @crvernon @arfon Therefore, we respectfully ask whether the issue could be reopened, allowing us the opportunity to address the newly raised concerns in a revised submission. We genuinely appreciate the constructive feedback provided so far and remain committed to improving the package further.

Thank you again for your time and consideration.
