### 2026-08-27 #8127 @LMKoesters
@crvernon Sounds good. We plan to fix the newly raised issues within the upcoming week - and have already fixed one of them. While we acknowledge that there still are oversights on our part, we put a lot of work into reworking the package from scratch (following Fabian's advice) and believe that the code is now a lot cleaner, more readable, and includes more thorough tests. The bugs mostly require minimal edits. The issue regarding the detection of autocorrelations not including factors and transforms was a conscious decision on our end (that, admittedly, should have been documented explicitly). However, we do believe that users will benefit from autocorrelations handling factors and transforms as well and will, therefore, implement this additional functionality.
I will update you as soon as the issues are fixed.
