### 2026-08-27 #8127 @fabian-s
@LMKoesters
With all due respect, the implementation (e.g. of variable selection, or weights and offsets) you re-submitted earlier remained *fundamentally* broken, which you did not figure out/ care to check by yourself.
Now you pretend as if the remaining and/or newly-introduced bugs and problems were just small doc fixes or defensible choices for default behavior -- they are not. Your software gives completely wrong answers, ignores/mishandles user input & implements non-sensical models.
I'm really taken aback by the carelessness it would take to advertise this for non-expert use without doing even minimal due dilligence. All of this should have been found and fixed by you / your coding agents.
In 2026, there's really no excuses for buggy code like this anymore.
Now @crvernon needs to waste another expert's time on this, what a shame...
