### 2026-08-27 #8127 @kevinkarbstein
@fabian-s I would like to intervene here. We are taking the technical concerns raised in the review seriously and will address them carefully in the revision. However, I think some of the language used in the review goes beyond constructive technical criticism and becomes unnecessarily personal.

Characterizing the authors' work as “careless,” questioning whether we “care to check” our implementation, or referring to the revision as something another expert has to “waste” their time on is not an appropriate way to discuss technical issues. We are happy to engage with detailed criticism of the code, models, defaults, documentation, or results, but we ask that such criticism remains factual and professional.

@LMKoesters and I are currently working through the issues raised in this round and will provide a substantive response to each of them. We would appreciate keeping the discussion focused on the technical aspects of the package rather than on assumptions about the authors' diligence or intentions.
