### 2026-08-27 #8127 @fabian-s
Fact is you tried to get a publication out of this pile of bugs again with an apparently minimal amount of effort and care and an obvious lack of regard for software quality or methodological standards.

My assessment of your diligence and competence is not an assumption anymore at this point, it's based on a lot of observation about your repeated failures to write correct software (as well as keeping to deadlines).

Instead of tone-policing me and *wasting* everybody's time once more with your sloppiness, maybe try doing *better work*...

Bad actors need to be called out even if it hurts their feelings, and you guys are bad actors:

1) If many JOSS submissions were as bad as yours and took up this much reviwer and editor time, JOSS could not go on.
2) If people like you who seem to be unable to write correct code insist on publishing bad statistics software, clueless users will do bad statistics with that software.

I do "professional" communication with people who demonstrate professionalism. That's definitely not you guys.
