### 2026-08-27 #8127 @kevinkarbstein
@fabian-s We are not trying to tone-police you or prevent you from raising serious concerns about the software. We fully accept that you should point out bugs, methodological problems, or cases where the implementation does not behave as intended, and we will investigate and address those issues.

However, there is a clear distinction between rigorous technical criticism and personal judgments about the authors. Statements that we are “sloppy,” “incompetent,” “bad actors,” or that we lack regard for software quality are not technical assessments of the package. They are personal characterizations, and we do not think they contribute constructively to the review.

We also do not agree with the characterization that we are deliberately trying to publish defective software or that we have a “lack of regard” for methodological standards. The purpose of this revision is precisely to address the issues identified during the review process and improve the package accordingly.

Regarding the time spent by reviewers and editors: we appreciate the time invested in reviewing the submission. At the same time, we do not think that frustration about the review process justifies attributing bad faith or incompetence to the authors (that is, us).

We suggest that we leave the personal assessments aside and focus the remainder of the review on concrete, reproducible technical and methodological issues. That will allow us to resolve the problems you have identified and allow the editor to make an informed decision about the suitability of the package for JOSS.
