### 2026-08-27 #8127 @crvernon
Yes, regardless of perspective we must keep things civil here.  We will work through this software review systematically and make a final decision based strictly on the findings.  I do appreciate the authors willingness to address the requested changes.  This is why I am bringing in another expert who specializes in R and statistics:  to independently see if what has been provided in response to comments is at the stage where minor changes can be made to prepare it for use by others or if there is still other work that needs to be done that would justify rejection - but with clear goals for the authors to address to help them in their objective of improving and maintaining this open source project.  Thank you.
